Update Invoice and Item models to calculate totals using float instead of integer.

+++ app/models/invoice.rb
@@ -16,7 +16,7 @@ class Invoice < ActiveRecord::Base
 
     self.items.each do |item|
       unless item.marked_for_destruction?
-        self.amount += (item.quantity.to_i * item.unit_cost.to_i) * ((100 - item.discount.to_i) / 100.0) + ((item.quantity.to_i * item.unit_cost.to_i) * 0.10)
+        self.amount += (item.quantity.to_f * item.unit_cost.to_f) * ((100 - item.discount.to_f) / 100.0) + ((item.quantity.to_f * item.unit_cost.to_f) * 0.10)
       end
     end
   end
