+++ app/models/item.rb
@@ -4,7 +4,7 @@ class Item < ActiveRecord::Base
 
   # Method to calculate total price of item after discount
   def total_price
-    (self.unit_cost.to_i * self.quantity.to_i) * ((100 - self.discount.to_i) / 100.0)
+    (self.unit_cost.to_f * self.quantity.to_f) * ((100 - self.discount.to_f) / 100.0)
   rescue NoMethodError => method
     0
   end
@@ -12,7 +12,7 @@ class Item < ActiveRecord::Base
   # Calculate invoice total with GST applied (10%)
   # TODO: update naming for clarity
   def calculate_gst
-    (self.unit_cost.to_i * self.quantity.to_i) * 0.10
+    (self.unit_cost.to_f * self.quantity.to_f) * 0.10
   rescue NoMethodError => method
     0
   end
